Widen object layer in/out/draw distance inputfields
Map dune_detail had 250000 set as stream out distance, but this
wasn't obvious in the layers window because there was only enough
space so it showed "2500" and the real value would only be obvious
when moving the caret in the inputfield. This caused an actual issue
so now the inputfields are a bit wider, which at least will make it
obvious if it has 5 digit values.